fix: corrections on module supplier and establishment

+++ backend/src/establishments/establishments.service.ts
@@ -1,5 +1,6 @@
 import {
   BadRequestException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
@@ -45,21 +46,10 @@ export class EstablishmentsService {
       await this.suppliersService.create({ supplierId: savedEstablishment.id})
     }
 
-
-    console.log('currentUser:', currentUser);
-    console.log('savedEstablishment.id:', savedEstablishment.id);
-
     // Lier le user à cet établissement
-    if (currentUser && currentUser.id) {
-      console.log('Tentative de mise à jour du user ID:', currentUser.id);
-      await this.userRepository.update(currentUser.id, {
+    await this.userRepository.update(currentUser.id, {
         establishmentId: savedEstablishment.id,
       });
-
-      console.log('User mis à jour');
-    } else {
-      console.log('❌ currentUser ou currentUser.userId manquant');
-    }
 
     return savedEstablishment;
   }
@@ -99,20 +89,30 @@ export class EstablishmentsService {
   async update(
     id: number,
     dto: UpdateEstablishmentDto,
+    currentUser: AuthenticatedUser,
   ): Promise<Establishment> {
     const establishment = await this.establishmentRepo.findOneBy({ id });
     if (!establishment) {
       throw new NotFoundException(`Establishment with ID ${id} not found`);
+    }
+    if (establishment.id !== currentUser.establishmentId) {
+      throw new ForbiddenException('You are not authorized to update this establishment.')
     }
     Object.assign(establishment, dto);
     return await this.establishmentRepo.save(establishment);
   }
 
   // Méthode pour supprimer un établissement
-  async remove(id: number): Promise<void> {
+  async remove(
+    id: number,
+    currentUser: AuthenticatedUser,
+  ): Promise<void> {
     const establishment = await this.establishmentRepo.findOneBy({ id });
     if (!establishment) {
       throw new NotFoundException(`Establishment with ID ${id} not found`);
+    }
+    if (establishment.id !== currentUser.establishmentId) {
+      throw new ForbiddenException('You are not authorized to delete this establishment.')
     }
     await this.establishmentRepo.remove(establishment);
   }
